# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Payment Server for save history of transactions and payments for testing.",
    "version": "0.0.0",
    "title": "Payment API",
    "contact": {
      "email": "ybrandao.d@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/yanBrandao/pay-api/blob/dev/LICENSE"
    }
  },
  "host": "localhost:3000",
  "basePath": "/api/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/transactions": {
      "post": {
        "tags": [
          "Transaction"
        ],
        "summary": "Add a new transaction",
        "operationId": "addTransaction",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "clientId",
            "type": "number",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "description": "Transaction object to add",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Transaction"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid Input"
          }
        }
      },
      "get": {
        "tags": [
          "Transaction"
        ],
        "summary": "List all Transactions",
        "operationId": "listAllTransaction",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Transaction"
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
...
    "CreditCard": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string",
          "example": "FULANO DE TAL"
        },
        "date_validation": {
          "type": "string",
          "format": "date",
          "example": "2022-12-08"
        },
        "card_number": {
          "type": "string",
          "example": "XXXX XXXX XXXX 9999"
        },
        "cvv": {
          "type": "string",
          "example": "999"
        }
      }
    },
    "Transaction": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "amount": {
          "type": "number",
          "example": 9999.99
        },
        "description": {
          "type": "string",
          "example": "Smartband XYZ 3.0"
        },
        "payment_method": {
          "type": "string",
          "example": "debit_card or credit_card"
        },
        "CreditCard": {
          "$ref": "#/definitions/CreditCard"
        }
      }
    },
...
  },
...
}